spo_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
notebooklm-py[browser]
playwright
